requirements.txt
SpeechRecognition
pyautogui
pyaudio
pystray
Pillow
vosk
pyinstaller
